ui: Major improvements to the library grid
Added a few book covers to use as placeholders for the library grid.
They are showing perfectly, although it's still not clear (to me) if
"hardcoding" width and height is a best practice or not. But, it works
and it's responsive!

I've also managed to get the text to clip and show an ellipsis, and to
drop a shadow from the book cover.
I removed the green background of the `BookItem`, as it's pretty much ok
now.