data: Hardcoded categories to be used
Until now, we had a function to retrieve the categories from a database.
This might not even be needed in the future, or it might be a little bit
overkill to have them be dynamic.

In any case, for now we'll have them hardcoded, since it's faster to
iterate that way.