ui: Added navbar for categories.
Below the `TopBar`, we now have a `NavigationMenu` from Radix-UI.
This new navigation bar has two buttons at each end, which allow for
smooth scrolling.
Needless to say, there are a bunch of things that are hardcoded for now.

Some changes were made to the `layout.tsx` and `data/categories.tsx`.
The former now calls the async function to retrieve the list of
categories (although they are hardcoded), and instantiates the new
`Categories` component; the latter now exposes the `Category` data type.